depth i8 -> i32

+++ src/engine/perft.rs
@@ -6,7 +6,7 @@ use crate::{board::board::Board, moves::movegenerator::generate_moves};
 
 #[allow(dead_code)]
 /// Counts and times the action of generating moves to a certain depth. Prints this information
-pub fn time_move_generation(board: &Board, depth: i8) {
+pub fn time_move_generation(board: &Board, depth: i32) {
     for i in 1..=depth {
         let start = Instant::now();
         print!("{}", count_moves(i, board));
@@ -16,7 +16,7 @@ pub fn time_move_generation(board: &Board, depth: i8) {
     }
 }
 
-pub fn multi_threaded_perft(board: Board, depth: i8) -> usize {
+pub fn multi_threaded_perft(board: Board, depth: i32) -> usize {
     let total = RwLock::new(0);
     let moves = generate_moves(&board);
     moves.into_vec().into_par_iter().for_each(|m| {
@@ -32,7 +32,7 @@ pub fn multi_threaded_perft(board: Board, depth: i8) -> usize {
     x
 }
 
-pub fn perft(board: Board, depth: i8) -> usize {
+pub fn perft(board: Board, depth: i32) -> usize {
     let mut total = 0;
     let moves = generate_moves(&board);
     for m in moves.iter() {
@@ -47,7 +47,7 @@ pub fn perft(board: Board, depth: i8) -> usize {
 }
 
 /// Recursively counts the number of moves down to a certain depth
-fn count_moves(depth: i8, board: &Board) -> usize {
+fn count_moves(depth: i32, board: &Board) -> usize {
     let mut count = 0;
     let moves = generate_moves(board);
     if depth == 1 {
